mrsbb/BackEnd/mrsbb-be/requirement.txt
Flask==2.2.2
Flask-Cors==3.0.10
wheel==0.40.0
rsa==4.7.2
pycryptodome==3.15.0
PyJWT==2.6.0
pandas==1.5.3
scikit-learn
pymongo==3.12.0
dnspython==2.3.0